Fix: overflow in update_ack

+++ src/kcp.rs
@@ -552,7 +552,7 @@ impl<Output> Kcp<Output> {
                 self.rx_srtt - rtt
             };
             self.rx_rttval = (3 * self.rx_rttval + delta) / 4;
-            self.rx_srtt = (7 * self.rx_srtt + rtt) / 8;
+            self.rx_srtt = ((7 * u64::from(self.rx_srtt) + u64::from(rtt)) / 8) as u32;
             if self.rx_srtt < 1 {
                 self.rx_srtt = 1;
             }
